+++ panther/request.py
@@ -1,10 +1,13 @@
+import logging
 from typing import Literal
 
 import orjson as json
 
 from panther._utils import read_multipart_form_data
 from panther.base_request import BaseRequest
-from panther.logger import logger
+
+
+logger = logging.getLogger('panther')
 
 
 class Request(BaseRequest):